extract cart dependencies from user bundle
